requirements.txt
asgiref==3.2.10
backcall==0.2.0
decorator==4.4.2
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
ipdb==0.13.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.7
ptyprocess==0.6.0
Pygments==2.7.1
PyJWT==1.7.1
python-dotenv==0.14.0
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.5
